refactor folder structure
