+++ client-app/src/app/layout/LoadingComponent.tsx
