+++ client-app/src/app/layout/NavBar.tsx
