+++ client-app/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./layout/App.tsx";
 import "semantic-ui-css/semantic.min.css";
+import App from "./app/layout/App.tsx";
 import "./layout/styles.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
